add tests

+++ src/node_info.rs
@@ -17,8 +17,7 @@ impl NodeInfo {
     pub fn new(seed: String, stellar_network: &Vec<u8>) -> NodeInfo {
         let key_pair = crypto::KeyPair::from_secret_seed(&seed).unwrap();
 
-        let mut network_id: [u8; 32i64 as usize] = Default::default();
-        // network_id.copy_from_slice(&network::Network::test_network().network_id()[..]);
+        let mut network_id: [u8; 32] = Default::default();
         network_id.copy_from_slice(&stellar_network[..]);
 
         NodeInfo {
